Matrixpilot_fbw - yet another attempt to process int to intx_t
git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_fbw@2322 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ MatrixPilot/main.c
@@ -23,7 +23,7 @@
 
 //	main program for testing the IMU.
 
-int main (void)
+int16_t main (void)
 {
 	udb_init() ;
 	dcm_init() ;
