+++ MatrixPilotFBW/autopilotCntrl.h
@@ -27,12 +27,12 @@
 // Transform demand positions and rotations into 
 void autopilotCntrl(void);
 
-void get_demand_rmat(int* prmat);
+void get_demand_rmat(int16_t* prmat);
 
 struct relative2D get_auto_rollDemand(void);
 struct relative2D get_auto_pitchDemand(void);
-int get_auto_rollRate(void);
-int get_auto_pitchRate(void);
+int16_t get_auto_rollRate(void);
+int16_t get_auto_pitchRate(void);
 
 //struct relative2D
